Massive update to the rewards onboarding screen :)
