--- Client/Frontend/Browser/Onboarding/OnboardingShieldsViewController.swift
@@ -1,31 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import UIKit
-import Shared
-import BraveShared
-
-class OnboardingShieldsViewController: OnboardingViewController {
-    
-    private var contentView: View {
-        return view as! View // swiftlint:disable:this force_cast
-    }
-    
-    override func loadView() {
-        view = View()
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        contentView.continueButton.addTarget(self, action: #selector(continueTapped), for: .touchUpInside)
-        contentView.skipButton.addTarget(self, action: #selector(skipTapped), for: .touchUpInside)
-    }
-    
-    override func continueTapped() {
-        Preferences.General.basicOnboardingProgress.value = OnboardingProgress.searchEngine.rawValue
-            
-        super.continueTapped()
-    }
-}
