+++ Client/Frontend/Browser/Onboarding/OnboardingWebViewController.swift
@@ -8,7 +8,14 @@ import Shared
 
 class OnboardingWebViewController: UIViewController, WKNavigationDelegate {
     
-    private let url = URL(string: "https://brave.com/terms-of-use/")
+    enum URLType {
+        case termsOfService
+        case privacyPolicy
+    }
+    
+    private let urlType: URLType
+    private let tosURL = URL(string: "https://brave.com/terms-of-use/")
+    private let ppURL = URL(string: "https://brave.com/privacy/browser/")
     private var helpers = [String: TabContentScript]()
     private var profile: Profile
     
@@ -44,8 +51,9 @@ class OnboardingWebViewController: UIViewController, WKNavigationDelegate {
         KVOs.forEach { webView.removeObserver(self, forKeyPath: $0.rawValue) }
     }
     
-    init(profile: Profile) {
+    init(profile: Profile, url: URLType) {
         self.profile = profile
+        self.urlType = url
         super.init(nibName: nil, bundle: nil)
     }
     
@@ -71,7 +79,15 @@ class OnboardingWebViewController: UIViewController, WKNavigationDelegate {
         
         webView.navigationDelegate = self
         setupScripts()
-        webView.load(PrivilegedRequest(url: url!) as URLRequest)
+        
+        switch urlType {
+        case .termsOfService:
+            guard let tosURL = tosURL else { return }
+            webView.load(PrivilegedRequest(url: tosURL) as URLRequest)
+        case .privacyPolicy:
+            guard let ppURL = ppURL else { return }
+            webView.load(PrivilegedRequest(url: ppURL) as URLRequest)
+        }
         
         toolbar.exitButton.addTarget(self, action: #selector(onExit), for: .touchUpInside)
         toolbar.backButton.addTarget(self, action: #selector(onBack), for: .touchUpInside)
@@ -90,8 +106,15 @@ class OnboardingWebViewController: UIViewController, WKNavigationDelegate {
 
         switch path {
         case .URL:
-            if url?.origin == webView.url?.origin {
-                toolbar.urlLabel.text = webView.url?.host
+            switch urlType {
+            case .termsOfService:
+                if tosURL?.origin == webView.url?.origin {
+                    toolbar.urlLabel.text = webView.url?.host
+                }
+            case .privacyPolicy:
+                if ppURL?.origin == webView.url?.origin {
+                    toolbar.urlLabel.text = webView.url?.host
+                }
             }
         case .canGoBack:
             updateBackForwardUI()
